pyproject.toml
[project]
name = "mymr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.17",
    "fastapi[standard]>=0.128.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]
